moved everything to central serial_eui_setter
Removed general serial_interface because it is out of scope for this ticket
