--- Projects/Applications/FreeRTOS/FreeRTOS_LoRaWAN/Core/App/Inc/credential_setter.h
@@ -1,10 +1,0 @@
-/*
- * credential_setter.h
- *
- *  Created on: May 15, 2025
- *      Author: robdr
- */
-#pragma once
-
-#include <string.h>
-#include "variable_credentials.h"
