+++ Projects/Applications/FreeRTOS/FreeRTOS_LoRaWAN/Core/App/Inc/serial_eui_setter.h
@@ -5,11 +5,9 @@
 #include "stm32wlxx_hal.h"
 #include "usart.h"
 #include "sys_app.h"
+#include "variable_credentials.h"
 
 #define RX_BUFFER_SIZE 10
-
-extern uint8_t rx_buffer[RX_BUFFER_SIZE];
-extern uint8_t rx_buffer_index;
 
 void Print_Rx_Buffer(void);
 void Add_To_Rx_Buffer(uint8_t *rxChar);
